+++ internal/controller/cluster_wide_quota.go
@@ -32,13 +32,13 @@ func (r *PaasReconciler) FetchAllPaasCapabilityResources(
 	ctx context.Context,
 	quota *quotav1.ClusterResourceQuota,
 	defaults map[corev1.ResourceName]resourcev1.Quantity,
-) (resources paasquota.QuotaLists, err error) {
+) (resources paasquota.Quotas, err error) {
 	capabilityName, err := ClusterWideCapabilityName(quota.Name)
 	if err != nil {
 		return resources, err
 	}
 	paas := &v1alpha1.Paas{}
-	resources = paasquota.NewQuotaLists()
+	resources = paasquota.NewQuotas()
 	for _, reference := range quota.OwnerReferences {
 		paasNamespacedName := types.NamespacedName{Name: reference.Name}
 		if reference.Kind != "Paas" || reference.APIVersion != v1alpha1.GroupVersion.String() {
@@ -68,6 +68,7 @@ func (r *PaasReconciler) UpdateClusterWideQuotaResources(
 	quota *quotav1.ClusterResourceQuota,
 ) (err error) {
 	var configCapability v1alpha1.ConfigCapability
+	var allPaasResources paasquota.Quotas
 	var allPaasResources paasquota.QuotaLists
 	if capabilityName, err := ClusterWideCapabilityName(quota.Name); err != nil {
 		return err
